[CCUBE-1522-1][JY] Remove unnecessary barrel imports. Shift regexes into common file.
