+++ src/translator/helper.ts
@@ -9,9 +9,9 @@ import {
     TElementMap,
 } from "src/context-providers/builder";
 import { ELEMENT_CONDITION_TYPES, SCHEMA_CONDITION_TYPES } from "src/data";
+import { PREFILL_ACTIONID_REGEX, PREFILL_PATH_REGEX } from "src/util";
 import { TextBasedField } from "./text-based-field";
 import { IPrefillConfig } from "./types";
-import { PREFILL_ACTIONID_REGEX, PREFILL_PATH_REGEX } from "src/yup-schemas";
 
 export const createPrefillObject = (elements: TElementMap) => {
     const prefill = Object.values(elements).reduce((acc, element) => {
