+++ src/util/index.ts
@@ -1,1 +1,2 @@
 export * from "./element-object-generator";
+export * from "./regexes";
