+++ src/util/regexes.ts
@@ -1,0 +1,3 @@
+export const ID_REGEX = /^[a-z]+(?:[A-Z0-9][a-z0-9]*)*(?:[-_][a-z0-9]+)*$/gm;
+export const PREFILL_ACTIONID_REGEX = /^[a-zA-Z0-9_-]+$/;
+export const PREFILL_PATH_REGEX = /^[a-zA-Z0-9._-]+$/;
