+++ src/yup-schemas/base-yup-schema-helper.ts
@@ -3,11 +3,8 @@ import {
     IFocusedElement,
     TElementMap,
 } from "src/context-providers";
+import { ID_REGEX, PREFILL_ACTIONID_REGEX, PREFILL_PATH_REGEX } from "src/util";
 import * as yup from "yup";
-
-const ID_REGEX = /^[a-z]+(?:[A-Z0-9][a-z0-9]*)*(?:[-_][a-z0-9]+)*$/gm;
-export const PREFILL_ACTIONID_REGEX = /^[a-zA-Z0-9_-]+$/;
-export const PREFILL_PATH_REGEX = /^[a-zA-Z0-9._-]+$/;
 
 export namespace BaseYupSchemaHelper {
     export const getBaseYupSchema = (
@@ -68,7 +65,4 @@ export namespace BaseYupSchemaHelper {
                 })
             ),
         });
-    export type TBaseSchema = yup.InferType<
-        ReturnType<typeof getBaseYupSchema>
-    >;
 }
